BUG: size_t introduced a bug in wrapping, removed by MeasurementMap
Added in Hans' MeasurementMap class.  Modified the printing function
to be a bit nicer and inline with SimpleITK's philosophy.

SIMPLEITK-78 #close

ENH:  Syncronized behavior of *Statistics

This creates a new generic measurement class
that is now used by both StatisticsImageFilter and
LabelStatisticsImageFilter.  Backwards compatibility
was maintained for StatisticsImageFilter.

The code in both filters is now mostly the same.

STYLE:  Const correctness was enforced on "Get" functions.

Conflicts:

	Code/Display/sitkShow.cxx